+++ CHANGELOG.rst
@@ -22,6 +22,8 @@ Added
 
 - Add ``get_servers`` and ``get_replicasets`` API
 
+- Add logging of configuration options on start and boot instance.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Changed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~